+++ FakeServer/Common/Config.cs
@@ -1,0 +1,8 @@
+﻿namespace FakeServer.Common
+{
+    public class Config
+    {
+        public const string ApiRoute = "api";
+        public const string AsyncRoute = "async";
+    }
+}